Bug fix, remove tss-react
Fix column sizing.  I've bounced back and forth between using native
html tables and styled divs with aria roles over the years, but at the
end of the day I've decided to stick with divs for this version.

In part because column sixing is so much easier to get working as I want
when it's implemented like this, and because I want to experiment with
sticky headers with a scroll bar that's bound to the table body, and
currently that's only possible if you avoid using the html table
controls.

General cleanup for mui v5, use sx where possible.